AC-743: Removed unused constant
